requirements.txt
pymongo
cassandra-driver
neo4j
tabulate
pyvis
faker
psycopg2